1，正在添加 ffmpeg 提取字幕和音频的功能 2，增加日志的详细程度

+++ internal/logic/sub_timeline_fixer/sub_timeline_fixer_helper.go
@@ -109,6 +109,10 @@ func (s SubTimelineFixerHelper) fixOneVideoSub(videoId string, videoRootPath str
 		if strings.Contains(exSubInfo.FileName, sub_timeline_fixer.FixMask) == false {
 			continue
 		}
+
+		if videoRootPath == "" {
+			continue
+		}
 		err = os.Remove(filepath.Join(videoRootPath, exSubInfo.FileName))
 		if err != nil {
 			return err
